==> 01rscripts/04ensemble/03_training_config/01_train_control_pretrain_stack.R <==
#*******************************************************************************
#*******************************************************************************
#######################    train control  ######################################
#*******************************************************************************
#*******************************************************************************

# config train control

#*******************************************************************************

#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
# define validation method
#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++

valid_method <- "repeatedcv"
valid_folds <- 2
valid_repeats <- 1

#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
# seeds
#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
set.seed(42)

#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
# define training control
#+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++

ctrl <- trainControl(
  method = valid_method, number = valid_folds, repeats = valid_repeats,
  summaryFunction = twoClassSummaryCustom,
  classProbs = TRUE,
  savePredictions = TRUE,
  allowParallel = TRUE,
  verboseIter = TRUE,
  selectionFunction = "best"
)
